Add files via upload
In this excercise the request was to be able to enstablish a connection between more devices, in this case the laptop-PT0 with IP 192.168.100.100/24 with the PC-PT-PC0 with IP 192.168.100.103/24,  belonging to the same IP net, and then between the laptop-PT0 with IP 192.168.100.100/24 and the laptop-PT2 with IP 192.168.200.100/24, belonging to different IP Net. In the first part to enstablish the connection I had to set the correct parameters, IP and subnetmask, and because they belong to the same IP Network 192.168.0.0/24 (Class C) I just had to make sure they were able to communicate, using the ping command. In the second part of the request, the two devices belonged to different IP Net, so to make them being able to communicate, i had to insert a Router Gateway. This device, (layer 3 ISO OSI) is able to put in contact devices from different IP Net. To make it happens, it need to be configured with certain parameters as the IP Gateway,. In this specific case i used the conventional IP Gateway 192.168.100.1/24 in the first IP net, and in the router port connected to it, and the conventional IP gateway 192.168.200.1/24 in the second group, and the coincident router's port. To enstablish a communication between them the first laptop (A) will send a packet together with some information as The IP Address of the other device (B) as destination in the datagram's header (level 3) and the router's MAC address as destination in the header's frame (level 2), while as sorgent, it will have his IP address (in lv3) and his MAC address (in lv2). When the packet will reach the router, it will check his routing table, and it will set as destination the MAC address of the B device, and as sorgent his MAC address. Then the packet will reach the device B and they will be able to comminicate. During the command ping the first and soeetimes the second packet too, could be lost, but  this is completly normal because the net need to settle down.